Implement first range of suggestions
