+++ solvers/circuitprocessing/circuitexecution/requirements.txt
@@ -1,3 +1,3 @@
-pytket
-pytket-qiskit
+pytket == 2.18.0
+pytket-qiskit == 0.77.0
 # pytket-qulacs
