+++ solvers/circuitprocessing/circuitoptimizing/requirements.txt
@@ -1,1 +1,1 @@
-pytket
+pytket == 2.18.0
