+++ src/main/java/edu/kit/provideq/toolbox/circuit/processing/solvers/MoveToMitigationSolver.java
@@ -1,7 +1,7 @@
-package edu.kit.provideq.toolbox.circuit.processing.solver;
+package edu.kit.provideq.toolbox.circuit.processing.solvers;
 
 import edu.kit.provideq.toolbox.Solution;
-import edu.kit.provideq.toolbox.circuit.processing.solver.mitigation.ErrorMitigationConfiguration;
+import edu.kit.provideq.toolbox.circuit.processing.solvers.mitigation.ErrorMitigationConfiguration;
 import edu.kit.provideq.toolbox.meta.SolvingProperties;
 import edu.kit.provideq.toolbox.meta.SubRoutineDefinition;
 import edu.kit.provideq.toolbox.meta.SubRoutineResolver;
