+++ src/main/resources/application.properties
@@ -125,7 +125,7 @@ venv.demonstrators.qiskit.molecule-energy=${name.demonstrators}_${name.demonstra
 name.circuitprocessing=circuitprocessing
 path.circuitprocessing=${name.solvers}/${name.circuitprocessing}
 name.circuitprocessing.circuitexecution=circuitexecution
-path.circuitprocessing.circuitexecution=${path.circuitprocessing}/${name.circuitprocessing.circuitexecution}/
+path.circuitprocessing.circuitexecution=${path.circuitprocessing}/${name.circuitprocessing.circuitexecution}/executor.py
 venv.circuitprocessing.circuitexecution=${name.solvers}_${name.circuitprocessing}_${name.circuitprocessing.circuitexecution}
 name.circuitprocessing.circuitoptimization=circuitoptimizing
 path.circuitprocessing.circuitoptimization=${path.circuitprocessing}/${name.circuitprocessing.circuitoptimization}/
