+++ src/test/java/edu/kit/provideq/toolbox/api/ExecutionSolverTest.java
@@ -1,10 +1,10 @@
 package edu.kit.provideq.toolbox.api;
 
-import static edu.kit.provideq.toolbox.circuit.processing.solver.executor.ExecutorConfiguration.EXECUTOR_CONFIG;
+import static edu.kit.provideq.toolbox.circuit.processing.solvers.executor.ExecutorConfiguration.EXECUTOR_CONFIG;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import edu.kit.provideq.toolbox.Solution;
-import edu.kit.provideq.toolbox.circuit.processing.solver.executor.ExecutionSolver;
+import edu.kit.provideq.toolbox.circuit.processing.solvers.executor.ExecutionSolver;
 import edu.kit.provideq.toolbox.meta.Problem;
 import edu.kit.provideq.toolbox.meta.ProblemManagerProvider;
 import java.time.Duration;
